Added Book service implementation and changed Book attribute bookIsbn to Long

+++ src/main/java/com/gustavo/sfrestlibrary/model/BookDTO.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 @Builder
 public class BookDTO {
     private UUID bookId;
-    private int bookIsbn;
+    private Long bookIsbn;
     private String bookName;
     private BookGenre bookGenre;
     private Integer bookVersion;
