+++ src/main/java/com/gustavo/sfrestlibrary/services/BookService.java
@@ -9,7 +9,8 @@ import java.util.UUID;
 
 public interface BookService {
     BookDTO createBook(BookDTO book);
-    Page<BookDTO> listAllBooks(int bookIsbn, String bookName, BookGenre bookgenre, Integer bookVersion, Integer pageNumber, Integer pageSize);
+    Page<BookDTO> listAllBooks(Long bookIsbn, String bookName, BookGenre bookgenre, Integer bookVersion,
+                               Integer pageNumber, Integer pageSize);
     Optional<BookDTO> getBookById(UUID bookId);
     Optional<BookDTO> updateBookById(UUID bookId, BookDTO book);
     Optional<BookDTO> patchBookById(UUID bookId, BookDTO book);
